+++ moor/test/schema_test.dart
@@ -37,7 +37,7 @@ void main() {
           'CREATE TABLE IF NOT EXISTS users '
           '(id INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT, '
           'name VARCHAR NOT NULL, '
-          'is_awesome BOOLEAN NOT NULL DEFAULT 1 CHECK (is_awesome in (0, 1)), '
+          'is_awesome INTEGER NOT NULL DEFAULT 1 CHECK (is_awesome in (0, 1)), '
           'profile_picture BLOB NOT NULL, '
           'creation_time INTEGER NOT NULL '
           "DEFAULT (strftime('%s', CURRENT_TIMESTAMP)));",
@@ -70,7 +70,7 @@ void main() {
           'CREATE TABLE IF NOT EXISTS users '
           '(id INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT, '
           'name VARCHAR NOT NULL, '
-          'is_awesome BOOLEAN NOT NULL DEFAULT 1 CHECK (is_awesome in (0, 1)), '
+          'is_awesome INTEGER NOT NULL DEFAULT 1 CHECK (is_awesome in (0, 1)), '
           'profile_picture BLOB NOT NULL, '
           'creation_time INTEGER NOT NULL '
           "DEFAULT (strftime('%s', CURRENT_TIMESTAMP)));",
@@ -88,7 +88,7 @@ void main() {
           .addColumn(db.users, db.users.isAwesome);
 
       verify(mockQueryExecutor.call('ALTER TABLE users ADD COLUMN '
-          'is_awesome BOOLEAN NOT NULL DEFAULT 1 CHECK (is_awesome in (0, 1));'));
+          'is_awesome INTEGER NOT NULL DEFAULT 1 CHECK (is_awesome in (0, 1));'));
     });
   });