+++ sqlparser/lib/src/analysis/schema/from_create_table.dart
@@ -2,6 +2,12 @@ part of '../analysis.dart';
 
 /// Reads the [Table] definition from a [CreateTableStatement].
 class SchemaFromCreateTable {
+  /// Whether we should provide additional type hints for nonstandard `BOOL`
+  /// and `DATETIME` columns.
+  final bool moorExtensions;
+
+  SchemaFromCreateTable({this.moorExtensions = false});
+
   Table read(CreateTableStatement stmt) {
     return Table(
       name: stmt.tableName,
@@ -13,10 +19,12 @@ class SchemaFromCreateTable {
   }
 
   TableColumn _readColumn(ColumnDefinition definition) {
-    final affinity = columnAffinity(definition.typeName);
+    final typeName = definition.typeName.toUpperCase();
+
+    final type = resolveColumnType(typeName);
     final nullable = !definition.constraints.any((c) => c is NotNull);
 
-    final resolvedType = ResolvedType(type: affinity, nullable: nullable);
+    final resolvedType = type.withNullable(nullable);
 
     return TableColumn(
       definition.columnName,
@@ -25,29 +33,46 @@ class SchemaFromCreateTable {
     );
   }
 
-  /// Looks up the correct column affinity for a declared type name with the
-  /// rules described here:
+  /// Resolves a column type via its typename, see the linked rules below.
+  /// Additionally, if [moorExtensions] are enabled, we support [IsBoolean] and
+  /// [IsDateTime] hints if the type name contains `BOOL` or `DATE`,
+  /// respectively.
   /// https://www.sqlite.org/datatype3.html#determination_of_column_affinity
-  @visibleForTesting
-  BasicType columnAffinity(String typeName) {
+  ResolvedType resolveColumnType(String typeName) {
     if (typeName == null) {
-      return BasicType.blob;
+      return const ResolvedType(type: BasicType.blob);
     }
 
     final upper = typeName.toUpperCase();
     if (upper.contains('INT')) {
-      return BasicType.int;
+      return const ResolvedType(type: BasicType.int);
     }
     if (upper.contains('CHAR') ||
         upper.contains('CLOB') ||
         upper.contains('TEXT')) {
-      return BasicType.text;
+      return const ResolvedType(type: BasicType.text);
     }
 
     if (upper.contains('BLOB')) {
-      return BasicType.blob;
+      return const ResolvedType(type: BasicType.blob);
     }
 
-    return BasicType.real;
+    if (moorExtensions) {
+      if (upper.contains('BOOL')) {
+        return const ResolvedType.bool();
+      }
+      if (upper.contains('DATE')) {
+        return const ResolvedType(
+            type: BasicType.int, hint: const IsDateTime());
+      }
+    }
+
+    return const ResolvedType(type: BasicType.real);
   }
+
+  /// Looks up the correct column affinity for a declared type name with the
+  /// rules described here:
+  /// https://www.sqlite.org/datatype3.html#determination_of_column_affinity
+  @visibleForTesting
+  BasicType columnAffinity(String typeName) => resolveColumnType(typeName).type;
 }